+++ Tests/src/TestExt/DbFuncMatrixCollection.cs
@@ -1,0 +1,6 @@
+using Xunit;
+
+namespace TestMooSQL.src;
+
+[CollectionDefinition("DbFuncMatrix", DisableParallelization = true)]
+public class DbFuncMatrixCollection;